refactor(media): moved types into own folder
