+++ src/trakt/index.ts
@@ -1,10 +1,10 @@
+export * from './media';
+
 export * from './certifications';
 export * from './countries';
 export * from './genres';
 export * from './languages';
 export * from './lists';
 export * from './networks';
-export * from './people';
-export * from './summary';
 export * from './traktTypes';
 export * from './users';
