+++ src/trakt/media/index.ts
@@ -1,0 +1,2 @@
+export * from './people';
+export * from './summary';
